chore(folder-structure): Move constants, hooks to shared domain
